feat: MSVC build support — compile WaveDB with Visual Studio 2022
Migrate the C library build from MSYS2 UCRT64 GCC to MSVC 2022,
resolving ABI incompatibility with Node.js node-gyp which uses MSVC.

Key changes:
- Convert GCC __atomic_* builtins to C11 atomic_* functions
- Add ATOMIC_TYPE(T) macro for C/C++ atomic compatibility
- Rewrite spinlock to use C11 atomics with C++ std::atomic fallback
- Create windows_compat.h ensuring winsock2.h before windows.h
- Create unistd_compat.h for POSIX function mappings on MSVC
- Update dirent_compat.h to use windows_compat.h
- Fix portable_endian.h — remove htonll/ntohll (WinSock2 provides them)
- Add /std:c11 /experimental:c11atomics MSVC compile flags
- Add _CRT_NONSTDC_NO_WARNINGS and gtest_force_shared_crt
- Guard encryption.c with WAVEDB_NO_OPENSSL for builds without OpenSSL
- Fix page_file.c pread/pwrite using _lseeki64 for 64-bit offsets
- Replace __thread with THREAD_LOCAL macro (__declspec(thread) on MSVC)
- Fix hashmap.h: typeof→__typeof__, zero-size arrays→[1], statement exprs
- Update all source/test files for MSVC compatibility
- Fix test files: std::thread for pthread, rename 'small' var (Windows keyword)

+++ src/Util/threadding.c
@@ -6,7 +6,7 @@
 #include <stdlib.h>
 #include <string.h>
 #if _WIN32
-#include <windows.h>
+#include "Util/windows_compat.h"
 #else
 #include <pthread.h>
 #include <unistd.h>
@@ -93,8 +93,8 @@ int platform_core_count() {
   GetSystemInfo(&sysinfo);
   return sysinfo.dwNumberOfProcessors;
 }
-int platform_self() {
-  return GetCurrentThreadId();
+uint64_t platform_self() {
+  return (uint64_t)GetCurrentThreadId();
 }
 #else
 void platform_lock(pthread_mutex_t* lock) {
